backend-ingest/requirements-dev.txt
# Tests y CI (no producción)
pytest>=8,<9
pytest-django>=4.8,<5
